pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aci-tool"
version = "1.1.0"
description = "Attacker Credibility Index — score ransomware groups on reliability, threat follow-through, and integrity"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Alex Kaariainen"},
]
dependencies = [
    "pandas>=2.2",
    "pyarrow>=16.0",
    "pydantic>=2.8",
    "python-dateutil>=2.9",
    "rapidfuzz>=3.9",
    "requests>=2.32",
    "tqdm>=4.66",
    "sentence-transformers>=2.7.0",
    "torch>=2.1.0",
    "numpy>=1.23",
    "scikit-learn>=1.3.0",
    "python-dotenv>=1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
]

[project.scripts]
aci = "aci_tool.cli:main"

[tool.setuptools.packages.find]
include = ["aci_tool*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
